initial version of rejester.Registry with tests and a stub for rejester.TaskMaster.
Registry offers a "lock" method for running several commands without
any other workers executing a command.  Note however that this does
NOT provide rollbacks.  Fortunately, most things that need guaranteed
atomicity can be implemented as Lua scripts.  Several Registry methods
are implemented via Lua scripts already.  These methods require that a
lock be acquired first.  This can be deceptive, because it appears to
suggest that the lock will provide a rollback, when it really does
not.  Implementing a rollback at the application level would be hard
or impossible in many cases.