Better behavior when jobs request more memory than we have.
Change the message when SingleWorker.run() doesn't get work to be less
programmery.  Change the message when TaskMaster.get_work() skips a work
spec due to low memory to be an info message, and to include the amount
of memory needed/available.  Add a "enough_memory" config option to
override this check.