Let a ForkWorker's child do more than one job.
We'd like the children to restart regularly to avoid Python memory
allocator issues, but the exit-fork-query loop takes a while, to the point
that it gets in the way pretty significantly if jobs are very short.  If
a child is allowed to run for 10 seconds running jobs in a loop until that
deadline passes, it shouldn't get itself into too much trouble.

Add rejester.workers.LoopWorker class to do this.