Use CBOR for value encoding, better preserve binary strings.
In particular, this allows work unit keys that contain \x00 to work
correctly.  There is no special handling of this via the CLI "rejester"
tool; commands like "rejester work_units" will print out unprintable
characters, and you have to pass them through your shell to get them
back in.  The output of "work_units" also changed to support this better
(if you had code that was trying to parse it, just create a TaskMaster
object).

This change may cause work units to not be returned in alphabetical
order; this was never an API guarantee.  This will invalidate any existing
rejester/Redis installations.